Refactor file names for MRCs
