+++ src/lib/miss_rate_curve/meson.build
@@ -1,7 +1,7 @@
-basic_miss_rate_curve_dep = declare_dependency(
+miss_rate_curve_dep = declare_dependency(
     link_with: library(
-        'basic_miss_rate_curve_lib',
-        'basic_miss_rate_curve.c',
+        'miss_rate_curve_lib',
+        'miss_rate_curve.c',
         include_directories: include_directories('include'),
         dependencies: [
             common_dep,
