+++ test/mrc_test/meson.build
@@ -19,7 +19,7 @@ fixed_size_shards_test_exe = executable(
     ],
     dependencies: [
         splitmix64_dep,
-        basic_miss_rate_curve_dep,
+        miss_rate_curve_dep,
         glib_dep,
         fixed_size_shards_dep,
         olken_dep,
@@ -35,7 +35,7 @@ mimir_test_exe = executable(
     ],
     dependencies: [
         glib_dep,
-        basic_miss_rate_curve_dep,
+        miss_rate_curve_dep,
         mimir_dep,
         mimir_buckets_dep,
         olken_dep,
@@ -52,7 +52,7 @@ quickmrc_test_exe = executable(
     dependencies: [
         glib_dep,
         basic_histogram_dep,
-        basic_miss_rate_curve_dep,
+        miss_rate_curve_dep,
         olken_dep,
         quickmrc_dep,
         zipfian_random_dep,
@@ -71,7 +71,7 @@ parda_fixed_rate_shards_test_exe = executable(
         parda_fixed_rate_shards_dep,
         splitmix64_dep,
         basic_histogram_dep,
-        basic_miss_rate_curve_dep,
+        miss_rate_curve_dep,
         zipfian_random_dep,
         olken_dep,
     ],
@@ -89,7 +89,7 @@ fixed_rate_shards_test_exe = executable(
         fixed_rate_shards_dep,
         splitmix64_dep,
         basic_histogram_dep,
-        basic_miss_rate_curve_dep,
+        miss_rate_curve_dep,
         zipfian_random_dep,
         olken_dep,
     ],
