+++ test/mrc_test/mimir_test.c
@@ -7,7 +7,7 @@
 #include "mimir/buckets.h"
 #include "mimir/mimir.h"
 #include "mimir/private_buckets.h"
-#include "miss_rate_curve/basic_miss_rate_curve.h"
+#include "miss_rate_curve/miss_rate_curve.h"
 #include "olken/olken.h"
 #include "random/zipfian_random.h"
 #include "test/mytester.h"
