+++ test/mrc_test/parda_fixed_rate_shards_test.c
@@ -5,7 +5,7 @@
 #include "arrays/array_size.h"
 #include "parda_shards/parda_fixed_rate_shards.h"
 #include "logger/logger.h"
-#include "miss_rate_curve/basic_miss_rate_curve.h"
+#include "miss_rate_curve/miss_rate_curve.h"
 #include "olken/olken.h"
 #include "parda.h"
 #include "random/zipfian_random.h"
